added nima dependencies and fixed subdirectory parsing
Also, self.importedFiles was including png files and that was messing with the data_generator so I added the glob thing to get the jpg files only. (we should add the other formats too i.e. JPG, JPEG)
Right now, predictions are made on all the files (images) in a directory, then it'll enter one of the subdirs in the current dir, output predictions for all the files in there, go to the next subdir, etc.