added training infrastructure
first off, I made line 205 operating-system-independent. it'll be important later.
I chose to make an unscored and scored list of images because this avoids having to do an O(n) search everytime we want a new random image (we'd have to search a main list that'd have flags for whether or not an image was already scored)