added means for loading and saving personal model in GUI
Clarification: the user is not doing the loading and saving, but we may want to consider that. When the user clicks Train, we'll check if a model already exisits (the user has done previous Training), if so we'll load that one, if not the user gets to use brand spankin' new model.  When they click Finish, we'll save the model.